feat(api) #7: fixes to get session loading from files and redis tests working

+++ test/integration/test_refresh_with_mocked_http.py
@@ -9,7 +9,8 @@ def redis_client():
     client = redis.Redis(host="localhost", port=6379, decode_responses=True)
     yield client
     keys = client.keys("*")
-    client.delete(*keys)
+    if keys:
+        client.delete(*keys)
 
 
 @pytest.mark.asyncio
